+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 ## 0.3.0
-* Added `S.rotate`, `S.scale` and `S.offset` as a style property
-* Changed style property from List<dynamic> to List<StyleItem> and gesture property from List<dynamic> to List<GestureItem>
+* Major rewrite. Changed the `List<StyleItem>` format to a custom `StyleClass`.Changed the `List<GestureItem>` format to a custom `GestureClass`
+* Improved `elevation()` style property
+* Added `animate()`, `rotate()`, `scale()` and `offset()` properties to the StyleClass
+* Enhanced preformance
+* Other minor tweaks
 
 ## 0.2.0
 * Added all gesture properties found in the GestureDetector widget
-* Added elevation style property containing boxShadow presets.
+* Added `elevation` style property containing boxShadow presets.
 * Fixed bug with S.align() and S.alignChild() style properties. They now work with `topLeft`, `topRight`, `bottomLeft` and `bottomRight`!
 * General code improvements and optimizing